requirements.txt
langchain_astradb
pandas
langchain_google_genai
fastapi 
uvicorn 
python-multipart
jinja2
python-dotenv 
langchain 
langchain_core 
langchain_groq
ipykernel
langchain-huggingface
sentence-transformers

-e . #it will help execute setup.py file and install the package in editable mode, allowing you to make changes to the code without reinstalling the package.